requirements.txt
deepspeech==0.4.1
Django==2.1.5
djangorestframework==3.9.1
numpy==1.16.1
pytz==2018.9
scipy==1.2.0
